storeimage moved to common
